test: replace failing MatcherCast with custom matcher
More recent gmock versions ensure that for MatcherCast<T>(U), T is a
base-class of U. This is the opposite of what we need, we know we have
an object of type IpcLogLineMessage but we get it as IpcMessage into the
matcher. There doesn't seem to be an easy way to avoid this
static_assert inside gmock.

To work around this for our tests, make a custom matcher instead and use
that.